% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat}
\alias{dat}
\title{RFI RNA-seq Data}
\format{The dataset has 11911 rows and 32 columns corresponding to 32 RNA-seq
  samples from 8 pigs, each pig has 4 RNA-seq samples measured repeatedly at
  0, 2, 6, 24 hours after LPS treatments.}
\usage{
dat
}
\description{
This is the RFI RNA-seq data motivating our paper.
}
\examples{
data(dat)
dim(dat)
head(dat)
}
\keyword{datasets}
