% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{design}
\alias{design}
\title{Reparameterized Design Matrix}
\format{A design matrix with 8 columns \code{Intercept, line2, time2, time6,
  time24, linetime2, linetime6, linetime24}.}
\usage{
design
}
\description{
This is a reparameterized  matrix from the original design matrix that
includes \code{time, line, time x line} such that each column of the
reparameterized design matrix corresponds to a test of interest such as
line main effect, \code{time2 - time0}, \code{time6 - time0}, \code{time24 -
time0}, and interaction effect of \code{timei - time0} with \code{line}. When
using this reparameterized design matrix in the analysis, for example,
regression coefficients corresponding to its second column \code{(line2)} represents
the \code{Line} main effect.
}
\examples{
data(design)
colnames(design)
head(design)
}
