% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{res}
\alias{res}
\title{Data Containing Results of Our Proposed Method Applying to RFI RNA-seq data}
\format{A list with 3 components \describe{
\item{ori.res}{a list consisting
 of 2 components: \code{v}, which is a voom output including several
 components; \code{newlm}, which is the output from \code{glsCAR1}}
 \item{pqvalue}{a list of 2 components \code{pv} and
 \code{qv}. \code{pv} is a data frame of pvalues for each test such as
 \code{line2
   time}; \code{qv} is a data frame of corresponding q-values of the above
 p-values, calculated using the method by Nettleton 2006.} }}
\usage{
res
}
\description{
This data set contains outputs of \code{\link[limma]{voom}},
\code{\link{glsCAR1}} and pqvalue of the tests for interested contrasts when analyzing the RFI RNA-seq
data using our proposed method.
}
\examples{
data(res)
names(res)
dim(res$ori.res$v)
colnames(res$ori.res$v)
colnames(res$ori.res$newlm)
colnames(res$pqvalue$pv)
colnames(res$pqvalue$qv)
}
\keyword{datasets}
