% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resSymm}
\alias{resSymm}
\title{Data Containing results when analyzing RFI using TC_Symm}
\format{
An object of class \code{list} of length 2.
}
\usage{
resSymm
}
\description{
This dataset contains the output of TC_Symm for 1000 genes
when applying the TC_Symm to the RFI dataset.
}
\examples{
data(resSymm)
names(resSymm)
dim(resSymm[[1]]$v)
colnames(resSymm[[1]]$newlm)
colnames(resSymm[[2]]$pv)
colnames(resSymm[[2]]$qv)
}
\keyword{datasets}
